yellow ballon

+++ src/RedBallon.pde
@@ -1,15 +1,9 @@
 class RedBallon extends Ballon {
-  RedBallon(float d) {
+  RedBallon(int d) {
     setBallonXLevel1(getBallonXLevel1() + d);
   }
-   RedBallon() {
-       
+   RedBallon() { 
   }
-  
-
-  
-  
-  
 
   void displayBallonLevel1(PImage RedBallonPhoto) {
     image(RedBallonPhoto, getBallonXLevel1(), getBallonYLevel1(), 80, 80);
